refactor(contract): move init, migrate and top-level views into api/lifecycle.rs
